chore: deprecate Markdown component
